hw07-atm02: actions moved to the dedicated "actions" package
