+++ hw07-atm2/src/main/java/ru/otus/vsh/hw07/api/actions/AtmGetValueAction.java
@@ -1,4 +1,4 @@
-package ru.otus.vsh.hw07.api;
+package ru.otus.vsh.hw07.api.actions;
 
 import ru.otus.vsh.hw07.impl.Department;
