+++ hw07-atm2/src/main/java/ru/otus/vsh/hw07/api/actions/AtmPutMoneyAction.java
@@ -1,5 +1,6 @@
-package ru.otus.vsh.hw07.api;
+package ru.otus.vsh.hw07.api.actions;
 
+import ru.otus.vsh.hw07.api.Banknote;
 import ru.otus.vsh.hw07.impl.Department;
 
 import javax.annotation.Nonnull;
